refactor(dashboard): handle metrics fetch with SWR
